backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
gitpython==3.1.43
pydantic==2.7.1
aiofiles==23.2.1
